requirements.txt
anyio==3.6.2
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
fhir.resources==7.0.0
fhirclient==4.1.0
h11==0.14.0
hapiclient==0.2.5
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
psycopg==3.1.9
psycopg2==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.14
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
